util: add log2f for Android
It is needed for nv50's new shader backend.  With this change, both u_math.h
and imports.h in core mesa define the same function.  I have to #undef log2f
here to avoid the conflict.  Not sure if there is a better way to deal with
the situation.